prune: remove dead code flagged by periphery
Declarations with zero references across Sources + Tests, verified with
periphery 3.7.4 (build+index scan) and corroborated by repo-wide ripgrep:

- TopNav view — superseded by FloatingRail; the only remaining reference
  was a stale doc comment. Its file is renamed to BurrowMark.swift, the one
  live type it still held (used by FloatingRail + PopupView).
- MoleClient.status(timeout:) — no callers; snapshots flow through the
  periodic producer path instead.
- EventHub.deregister(_:) — dead public counterpart; the live eviction path
  is the private deregisterID(_:).
- Brand.rounded(_:_:) — unused font helper.
- MenuBarSnapshot.has(_:) — no callers.

Also fixes the now-inaccurate SettingsView navigation comment.

Kept intentionally (periphery false positives): MoleStatus Codable mirror
and SMC kernel-struct fields (memory-layout-critical), Brand palette tokens,
and Telemetry bytes/count/seconds bucket helpers (documented privacy tooling).
